requirements.txt
langgraph==0.3.21
langgraph-prebuilt==0.1.7
streamlit==1.32.0
groq==0.4.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
tweepy==4.14.0
linkedin-api==2.0.3
facebook-sdk==3.1.0
python-linkedin==4.1
pandas==2.2.1
boto3==1.34.69
pytest==8.0.2
pytest-mock==3.12.0
pytest-cov==4.1.0
moto==4.2.7
pytest-asyncio==0.23.5 